Move more logic into dbt service module
